carga inicial a rama Diego
Se carga todo excepto los algoritmos entrenados y sus test . notebook individuales
