+++ references/Diccionario de variables.txt
@@ -1,0 +1,20 @@
+Diccionario de Variables
+A continuación se brinda el listado de variables utilizadas en el proyecto:
+ESTIMACIONES (Provincia de BsAs 1987-2022.csv)
+-Cultivo: Tipos de cultivos
+-Campaña: Periodo de tiempo de duración del rendimiento
+-Provincia: División administrativa en que se organiza el Estado
+-Departamento: Los Departamentos y partidos son las subdivisiones de segundo orden de una provincia
+-idprovincia: número asignado a cada Provincia
+-iddepartamento: número asignado a cada Departamento
+-Sup.Sembrada: Superficie en héctarea sembrada
+-Sup.Cosechada: Superficie en héctarea cosechada
+-Producción: Toneladas de producción
+-Rendimiento: El rendimiento se expresa kg/héctarea
+
+192602 TRES ARROYOS 1988-2022.xlsx / 192602 JUNIN AERO 1988-2022.xlsx
+-Fecha: dd/mm/aaaa
+-Temperatura máxima: mediciones en centrigrados
+-Temperatura media:  mediciones en centrigrados
+-Temperatura minima: mediciones en centrigrados
+-Precipitación: mediciones en milimetros
