--- setup.py
@@ -1,10 +1,0 @@
-from setuptools import find_packages, setup
-
-setup(
-    name='src',
-    packages=find_packages(),
-    version='0.1.0',
-    description='A short description of the project.',
-    author='Miguel',
-    license='MIT',
-)
