--- tox.ini
@@ -1,3 +1,0 @@
-[flake8]
-max-line-length = 79
-max-complexity = 10
